+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+#### v0.3.0
+- breaking: every event has multiple listeners (previously only on('message') allowed multiple)
+- added: join/leave/error event types
+- added: .remove('event-name', listener) to remove listeners
 #### v0.2.3
 - fix type hinting on listeners
 - improve type hinting on send/push